Ensure "endstream" marker is always preceded by a newline
When a stream consists of binary data (e.g. has the FlateDecode filter applied), the final byte might possibly be a Line Feed or a Carriage Return character which will trigger the conditional to skip adding the newline before the EndStream marker.  This in turn will mean that the number of bytes between the "stream" and "endstream" does not match the "Length".